+++ include/base/OpenMCProblemBase.h
@@ -32,14 +32,25 @@ public:
 
   static InputParameters validParams();
 
+  virtual ~OpenMCProblemBase() override;
+
   /// Run a k-eigenvalue OpenMC simulation
   void externalSolve() override;
+
+  /// Import temperature and density from a properties.h5 file
+  void importProperties() const;
 
   /**
    * Type definition for storing the relevant aspects of the OpenMC geometry; the first
    * value is the cell index, while the second is the cell instance.
    */
   typedef std::pair<int32_t, int32_t> cellInfo;
+
+  /**
+   * Get the density conversion factor (multiplicative factor)
+   * @return density conversion factor from kg/m3 to g/cm3
+   */
+  const Real & densityConversionFactor() const { return _density_conversion_factor; }
 
   /**
    * Get the number of particles used in the current Monte Carlo calculation
@@ -107,6 +118,49 @@ public:
    */
   unsigned int globalElemID(const unsigned int & id) const { return _local_to_global_elem[id]; }
 
+  /**
+   * Set the cell temperature, and print helpful error message if a failure occurs; this sets
+   * the temperature for the id + instance, which could be one of N cells filling the 'cell_info'
+   * parent cell (which is what we actually use for error printing)
+   * @param[in] id cell ID
+   * @param[in] instance cell instance
+   * @param[in] T temperature
+   * @param[in] cell_info cell info for which we are setting interior temperature, for error printing
+   */
+  virtual void setCellTemperature(const int32_t & id, const int32_t & instance, const Real & T,
+    const cellInfo & cell_info) const;
+
+  /**
+   * Set the cell density, and print helpful error message if a failure occurs
+   * @param[in] density density
+   * @param[in] cell_info cell info for which we are setting the density
+   */
+  virtual void setCellDensity(const Real & density, const cellInfo & cell_info) const;
+
+  /**
+   * Get a descriptive, formatted, string describing a cell
+   * @param[in] cell_info cell index, instance pair
+   * @return descriptive string describing cell
+   */
+  virtual std::string printCell(const cellInfo & cell_info) const;
+
+  /**
+   * Get the fill of an OpenMC cell
+   * @param[in] cell_info cell ID, instance
+   * @param[out] fill_type fill type of the cell, one of MATERIAL, UNIVERSE, or LATTICE
+   * @return indices of material fills
+   */
+  virtual std::vector<int32_t> cellFill(const cellInfo & cell_info, int & fill_type) const;
+
+  /**
+   * Whether a cell contains any fissile materials; for now, this simply returns true for
+   * cells filled by universes or lattices because we have yet to implement something more
+   * sophisticated that recurses down into all the fills
+   * @param[in] cell_info cell ID, instance
+   * @return whether cell contains fissile material
+   */
+  virtual bool cellHasFissileMaterials(const cellInfo & cell_info) const;
+
 protected:
   /**
    * Set an auxiliary elemental variable to a specified value
@@ -162,6 +216,20 @@ protected:
   /// Directory where OpenMC output files are written
   std::string _path_output;
 
+  /**
+   * Number of digits to use to display the cell ID for diagnostic messages; this is
+   * estimated conservatively based on the total number of cells, even though there
+   * may be distributed cells such that the maximum cell ID is far smaller than the
+   * total number of cells.
+   */
+  const int _n_cell_digits;
+
   /// Mapping from local element indices to global element indices for this rank
   std::vector<unsigned int> _local_to_global_elem;
+
+  /// Conversion unit to transfer between kg/m3 and g/cm3
+  static constexpr Real _density_conversion_factor{0.001};
+
+  /// ID used by OpenMC to indicate that a material fill is VOID
+  static constexpr int MATERIAL_VOID{-1};
 };
