+++ src/base/OpenMCCellAverageProblem.C
@@ -33,7 +33,6 @@
 #include "openmc/cell.h"
 #include "openmc/constants.h"
 #include "openmc/error.h"
-#include "openmc/material.h"
 #include "openmc/particle.h"
 #include "openmc/geometry.h"
 #include "openmc/geometry_aux.h"
@@ -235,7 +234,6 @@ OpenMCCellAverageProblem::OpenMCCellAverageProblem(const InputParameters & param
     _has_fluid_blocks(params.isParamSetByUser("fluid_blocks")),
     _has_solid_blocks(params.isParamSetByUser("solid_blocks")),
     _needs_global_tally(_check_tally_sum || _normalize_by_global),
-    _n_cell_digits(digits(openmc::model::cells.size())),
     _using_default_tally_blocks(_tally_type == tally::cell && _single_coord_level &&
                                 !isParamValid("tally_blocks")),
     _total_n_particles(0),
@@ -455,8 +453,6 @@ OpenMCCellAverageProblem::getTallyTriggerParameters(const InputParameters & para
     checkUnusedParam(parameters, "batch_interval", "not using triggers");
   }
 }
-
-OpenMCCellAverageProblem::~OpenMCCellAverageProblem() { openmc_finalize(); }
 
 template <typename T>
 void
@@ -679,12 +675,6 @@ OpenMCCellAverageProblem::readBlockParameters(const std::string name,
         mooseError("Block " + Moose::stringify(b) + " specified in '" + name +
                    "_blocks' not found in mesh!");
   }
-}
-
-int
-OpenMCCellAverageProblem::digits(const int & number) const
-{
-  return std::to_string(number).length();
 }
 
 const coupling::CouplingFields
@@ -1012,20 +1002,6 @@ OpenMCCellAverageProblem::getMaterialFills()
     _console << "\nMaterials in each OpenMC fluid cell:" << std::endl;
     vt.print(_console);
   }
-}
-
-std::string
-OpenMCCellAverageProblem::printCell(const cellInfo & cell_info) const
-{
-  int32_t id = cellID(cell_info.first);
-
-  std::stringstream msg;
-  msg << "id " << std::setw(_n_cell_digits) << Moose::stringify(id) << ", instance "
-      << std::setw(_n_cell_digits) << Moose::stringify(cell_info.second) << " (of "
-      << std::setw(_n_cell_digits)
-      << Moose::stringify(openmc::model::cells[cell_info.first]->n_instances_) << ")";
-
-  return msg.str();
 }
 
 void
@@ -2002,17 +1978,8 @@ OpenMCCellAverageProblem::sendTemperatureToOpenMC()
     auto contained_cells = _cell_to_contained_material_cells[cell_info];
     for (const auto & contained : contained_cells)
     {
-      auto id = contained.first;
-
       for (const auto & instance : contained.second)
-      {
-        int err = openmc_cell_set_temperature(id, average_temp, &instance, false);
-
-        if (err)
-          mooseError("In attempting to set cell " + printCell(cell_info) + " to temperature " +
-                     Moose::stringify(average_temp) + " (K), OpenMC reported:\n\n" +
-                     std::string(openmc_err_msg));
-      }
+        setCellTemperature(contained.first, instance, average_temp, cell_info);
     }
   }
 
@@ -2055,43 +2022,11 @@ OpenMCCellAverageProblem::sendDensityToOpenMC()
     minimum = std::min(minimum, average_density);
     maximum = std::max(maximum, average_density);
 
-    // OpenMC technically allows a density of >= 0.0, but we can impose a tighter
-    // check here with a better error message than the Excepts() in material->set_density
-    // because it could be a very common mistake to forget to set an initial condition
-    // for density if OpenMC runs first
-    if (average_density <= 0.0)
-      mooseError("Densities less than or equal to zero cannot be set in the OpenMC model!\n cell " +
-                 printCell(cell_info) + " set to density " + Moose::stringify(average_density) +
-                 " (kg/m3)");
-
     if (_verbose)
       _console << "Setting cell " << printCell(cell_info) << " to density (kg/m3): " << std::setw(4)
                << average_density << std::endl;
 
-    int fill_type;
-    std::vector<int32_t> material_indices = cellFill(cell_info, fill_type);
-
-    // throw a special error if the cell is void, because the OpenMC error isn't very
-    // clear what the mistake is
-    if (material_indices[0] == MATERIAL_VOID)
-      mooseError("Cannot set density for cell " + printCell(cell_info) +
-                 " because this cell is void (vacuum)!");
-
-    if (fill_type != static_cast<int>(openmc::Fill::MATERIAL))
-      mooseError(
-          "Density transfer does not currently support cells filled with universes or lattices!");
-
-    // Multiply density by 0.001 to convert from kg/m3 (the units assumed in the 'density'
-    // auxvariable as well as the MOOSE fluid properties module) to g/cm3
-    const char * units = "g/cc";
-    int err = openmc_material_set_density(
-        material_indices[cell_info.second], average_density * _density_conversion_factor, units);
-
-    if (err)
-      mooseError("In attempting to set material with index " +
-                 Moose::stringify(material_indices[cell_info.second]) + " to density " +
-                 Moose::stringify(average_density) + " (kg/m3), OpenMC reported:\n\n" +
-                 std::string(openmc_err_msg));
+    setCellDensity(average_density, cell_info);
   }
 
   if (!_verbose)
@@ -2443,14 +2378,7 @@ OpenMCCellAverageProblem::syncSolutions(ExternalProblem::Direction direction)
         {
           case coupling::hdf5:
           {
-            _console << "Reading temperature and density from properties.h5" << std::endl;
-
-            int err = openmc_properties_import("properties.h5");
-            if (err)
-              mooseError("In attempting to load temperature and density from a properties.h5 file, "
-                         "OpenMC reported:\n\n" +
-                         std::string(openmc_err_msg));
-
+            importProperties();
             return;
           }
           case coupling::moose:
@@ -2558,51 +2486,6 @@ OpenMCCellAverageProblem::checkTallySum() const
 
     mooseError(msg.str());
   }
-}
-
-std::vector<int32_t>
-OpenMCCellAverageProblem::cellFill(const cellInfo & cell_info, int & fill_type) const
-{
-  fill_type = static_cast<int>(openmc::Fill::MATERIAL);
-  int32_t * materials = nullptr;
-  int n_materials = 0;
-
-  int err = openmc_cell_get_fill(cell_info.first, &fill_type, &materials, &n_materials);
-
-  if (err)
-    mooseError("In attempting to get fill of cell " + printCell(cell_info) +
-               ", OpenMC reported:\n\n" + std::string(openmc_err_msg));
-
-  std::vector<int32_t> material_indices;
-  material_indices.assign(materials, materials + n_materials);
-  return material_indices;
-}
-
-bool
-OpenMCCellAverageProblem::cellHasFissileMaterials(const cellInfo & cell_info) const
-{
-  int fill_type;
-  std::vector<int32_t> material_indices = cellFill(cell_info, fill_type);
-
-  // TODO: for cells with non-material fills, we need to implement something that recurses
-  // into the cell/universe fills to see if there's anything fissile; until then, just assume
-  // that the cell has something fissile
-  if (fill_type != static_cast<int>(openmc::Fill::MATERIAL))
-    return true;
-
-  // for each material fill, check whether it is fissionable
-  for (const auto & index : material_indices)
-  {
-    // We know void cells certainly aren't fissionable; if not void, check if fissionable
-    if (index != MATERIAL_VOID)
-    {
-      const auto & material = openmc::model::materials[index];
-      if (material->fissionable_)
-        return true;
-    }
-  }
-
-  return false;
 }
 
 void
