+++ src/base/OpenMCProblem.C
@@ -138,8 +138,6 @@ OpenMCProblem::OpenMCProblem(const InputParameters & params)
   setupTallies();
 }
 
-OpenMCProblem::~OpenMCProblem() { openmc_finalize(); }
-
 void
 OpenMCProblem::fillCenters()
 {
